requirements.txt
flask==2.3.3
pillow==10.2.0
gunicorn==21.2.0